CovidAPP-main/AppCovid-aplicacionfuncional/requirements.txt
asgiref==3.5.0
base45==0.4.3
cbor2==5.4.2
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.10
colorama==0.4.4
cryptography==36.0.1
Deprecated==1.2.13
Django==4.0.1
gcloud==0.18.3
googleapis-common-protos==1.54.0
httplib2==0.20.2
idna==3.3
jwcrypto==1.0
numpy==1.22.1
oauth2client==4.1.3
opencv-python==4.5.5.62
Pillow==9.0.0
protobuf==3.19.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pycryptodome==3.13.0
PyMySQL==1.0.2
pyparsing==3.0.7
Pyrebase4==4.5.0
python-jwt==3.3.2
qrcode==7.3.1
requests==2.27.1
requests-toolbelt==0.9.1
rsa==4.8
six==1.16.0
sqlparse==0.4.2
tzdata==2021.5
urllib3==1.26.8
wrapt==1.13.3
